opt: streamline pool implementation to reduce duplicated code
