+++ .github/workflows/test.yml
@@ -55,7 +55,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        go: [1.16, 1.23]
+        go: [1.18, 1.23]
         os: [ubuntu-latest, macos-latest, windows-latest]
     name: Go ${{ matrix.go }} @ ${{ matrix.os }}
     runs-on: ${{ matrix.os}}
