+++ examples/main.go
@@ -33,7 +33,7 @@ import (
 
 var sum int32
 
-func myFunc(i interface{}) {
+func myFunc(i any) {
 	n := i.(int32)
 	atomic.AddInt32(&sum, n)
 	fmt.Printf("run with %d\n", n)
@@ -65,7 +65,7 @@ func main() {
 
 	// Use the pool with a function,
 	// set 10 to the capacity of goroutine pool and 1 second for expired duration.
-	p, _ := ants.NewPoolWithFunc(10, func(i interface{}) {
+	p, _ := ants.NewPoolWithFunc(10, func(i any) {
 		myFunc(i)
 		wg.Done()
 	})
@@ -96,7 +96,7 @@ func main() {
 	fmt.Printf("finish all tasks.\n")
 
 	// Use the MultiPoolFunc and set the capacity of 10 goroutine pools to (runTimes/10).
-	mpf, _ := ants.NewMultiPoolWithFunc(10, runTimes/10, func(i interface{}) {
+	mpf, _ := ants.NewMultiPoolWithFunc(10, runTimes/10, func(i any) {
 		myFunc(i)
 		wg.Done()
 	}, ants.LeastTasks)
