+++ go.mod
@@ -1,8 +1,14 @@
 module github.com/panjf2000/ants/v2
 
-go 1.16
+go 1.18
 
 require (
 	github.com/stretchr/testify v1.8.2
 	golang.org/x/sync v0.3.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
