+++ pool.go
@@ -22,203 +22,13 @@
 
 package ants
 
-import (
-	"context"
-	"sync"
-	"sync/atomic"
-	"time"
-
-	syncx "github.com/panjf2000/ants/v2/pkg/sync"
-)
-
-type poolCommon struct {
-	// capacity of the pool, a negative value means that the capacity of pool is limitless, an infinite pool is used to
-	// avoid potential issue of endless blocking caused by nested usage of a pool: submitting a task to pool
-	// which submits a new task to the same pool.
-	capacity int32
-
-	// running is the number of the currently running goroutines.
-	running int32
-
-	// lock for protecting the worker queue.
-	lock sync.Locker
-
-	// workers is a slice that store the available workers.
-	workers workerQueue
-
-	// state is used to notice the pool to closed itself.
-	state int32
-
-	// cond for waiting to get an idle worker.
-	cond *sync.Cond
-
-	// done is used to indicate that all workers are done.
-	allDone chan struct{}
-	// once is used to make sure the pool is closed just once.
-	once *sync.Once
-
-	// workerCache speeds up the obtainment of a usable worker in function:retrieveWorker.
-	workerCache sync.Pool
-
-	// waiting is the number of goroutines already been blocked on pool.Submit(), protected by pool.lock
-	waiting int32
-
-	purgeDone int32
-	purgeCtx  context.Context
-	stopPurge context.CancelFunc
-
-	ticktockDone int32
-	ticktockCtx  context.Context
-	stopTicktock context.CancelFunc
-
-	now atomic.Value
-
-	options *Options
-}
-
-// Pool accepts the tasks and process them concurrently,
-// it limits the total of goroutines to a given number by recycling goroutines.
+// Pool is a goroutine pool that limits and recycles a mass of goroutines.
+// The pool capacity can be fixed or unlimited.
 type Pool struct {
-	poolCommon
+	*poolCommon
 }
 
-// purgeStaleWorkers clears stale workers periodically, it runs in an individual goroutine, as a scavenger.
-func (p *Pool) purgeStaleWorkers() {
-	ticker := time.NewTicker(p.options.ExpiryDuration)
-
-	defer func() {
-		ticker.Stop()
-		atomic.StoreInt32(&p.purgeDone, 1)
-	}()
-
-	purgeCtx := p.purgeCtx // copy to the local variable to avoid race from Reboot()
-	for {
-		select {
-		case <-purgeCtx.Done():
-			return
-		case <-ticker.C:
-		}
-
-		if p.IsClosed() {
-			break
-		}
-
-		var isDormant bool
-		p.lock.Lock()
-		staleWorkers := p.workers.refresh(p.options.ExpiryDuration)
-		n := p.Running()
-		isDormant = n == 0 || n == len(staleWorkers)
-		p.lock.Unlock()
-
-		// Clean up the stale workers.
-		for i := range staleWorkers {
-			staleWorkers[i].finish()
-			staleWorkers[i] = nil
-		}
-
-		// There might be a situation where all workers have been cleaned up (no worker is running),
-		// while some invokers still are stuck in p.cond.Wait(), then we need to awake those invokers.
-		if isDormant && p.Waiting() > 0 {
-			p.cond.Broadcast()
-		}
-	}
-}
-
-// ticktock is a goroutine that updates the current time in the pool regularly.
-func (p *Pool) ticktock() {
-	ticker := time.NewTicker(nowTimeUpdateInterval)
-	defer func() {
-		ticker.Stop()
-		atomic.StoreInt32(&p.ticktockDone, 1)
-	}()
-
-	ticktockCtx := p.ticktockCtx // copy to the local variable to avoid race from Reboot()
-	for {
-		select {
-		case <-ticktockCtx.Done():
-			return
-		case <-ticker.C:
-		}
-
-		if p.IsClosed() {
-			break
-		}
-
-		p.now.Store(time.Now())
-	}
-}
-
-func (p *Pool) goPurge() {
-	if p.options.DisablePurge {
-		return
-	}
-
-	// Start a goroutine to clean up expired workers periodically.
-	p.purgeCtx, p.stopPurge = context.WithCancel(context.Background())
-	go p.purgeStaleWorkers()
-}
-
-func (p *Pool) goTicktock() {
-	p.now.Store(time.Now())
-	p.ticktockCtx, p.stopTicktock = context.WithCancel(context.Background())
-	go p.ticktock()
-}
-
-func (p *Pool) nowTime() time.Time {
-	return p.now.Load().(time.Time)
-}
-
-// NewPool instantiates a Pool with customized options.
-func NewPool(size int, options ...Option) (*Pool, error) {
-	if size <= 0 {
-		size = -1
-	}
-
-	opts := loadOptions(options...)
-
-	if !opts.DisablePurge {
-		if expiry := opts.ExpiryDuration; expiry < 0 {
-			return nil, ErrInvalidPoolExpiry
-		} else if expiry == 0 {
-			opts.ExpiryDuration = DefaultCleanIntervalTime
-		}
-	}
-
-	if opts.Logger == nil {
-		opts.Logger = defaultLogger
-	}
-
-	p := &Pool{poolCommon: poolCommon{
-		capacity: int32(size),
-		allDone:  make(chan struct{}),
-		lock:     syncx.NewSpinLock(),
-		once:     &sync.Once{},
-		options:  opts,
-	}}
-	p.workerCache.New = func() interface{} {
-		return &goWorker{
-			pool: p,
-			task: make(chan func(), workerChanCap),
-		}
-	}
-	if p.options.PreAlloc {
-		if size == -1 {
-			return nil, ErrInvalidPreAllocSize
-		}
-		p.workers = newWorkerQueue(queueTypeLoopQueue, size)
-	} else {
-		p.workers = newWorkerQueue(queueTypeStack, 0)
-	}
-
-	p.cond = sync.NewCond(p.lock)
-
-	p.goPurge()
-	p.goTicktock()
-
-	return p, nil
-}
-
-// Submit submits a task to this pool.
+// Submit submits a task to the pool.
 //
 // Note that you are allowed to call Pool.Submit() from the current Pool.Submit(),
 // but what calls for special attention is that you will get blocked with the last
@@ -236,198 +46,20 @@ func (p *Pool) Submit(task func()) error {
 	return err
 }
 
-// Running returns the number of workers currently running.
-func (p *Pool) Running() int {
-	return int(atomic.LoadInt32(&p.running))
-}
-
-// Free returns the number of available workers, -1 indicates this pool is unlimited.
-func (p *Pool) Free() int {
-	c := p.Cap()
-	if c < 0 {
-		return -1
-	}
-	return c - p.Running()
-}
-
-// Waiting returns the number of tasks waiting to be executed.
-func (p *Pool) Waiting() int {
-	return int(atomic.LoadInt32(&p.waiting))
-}
-
-// Cap returns the capacity of this pool.
-func (p *Pool) Cap() int {
-	return int(atomic.LoadInt32(&p.capacity))
-}
-
-// Tune changes the capacity of this pool, note that it is noneffective to the infinite or pre-allocation pool.
-func (p *Pool) Tune(size int) {
-	capacity := p.Cap()
-	if capacity == -1 || size <= 0 || size == capacity || p.options.PreAlloc {
-		return
-	}
-	atomic.StoreInt32(&p.capacity, int32(size))
-	if size > capacity {
-		if size-capacity == 1 {
-			p.cond.Signal()
-			return
-		}
-		p.cond.Broadcast()
-	}
-}
-
-// IsClosed indicates whether the pool is closed.
-func (p *Pool) IsClosed() bool {
-	return atomic.LoadInt32(&p.state) == CLOSED
-}
-
-// Release closes this pool and releases the worker queue.
-func (p *Pool) Release() {
-	if !atomic.CompareAndSwapInt32(&p.state, OPENED, CLOSED) {
-		return
+// NewPool instantiates a Pool with customized options.
+func NewPool(size int, options ...Option) (*Pool, error) {
+	pc, err := newPool(size, options...)
+	if err != nil {
+		return nil, err
 	}
 
-	if p.stopPurge != nil {
-		p.stopPurge()
-		p.stopPurge = nil
-	}
-	if p.stopTicktock != nil {
-		p.stopTicktock()
-		p.stopTicktock = nil
-	}
-
-	p.lock.Lock()
-	p.workers.reset()
-	p.lock.Unlock()
-	// There might be some callers waiting in retrieveWorker(), so we need to wake them up to prevent
-	// those callers blocking infinitely.
-	p.cond.Broadcast()
-}
-
-// ReleaseTimeout is like Release but with a timeout, it waits all workers to exit before timing out.
-func (p *Pool) ReleaseTimeout(timeout time.Duration) error {
-	if p.IsClosed() || (!p.options.DisablePurge && p.stopPurge == nil) || p.stopTicktock == nil {
-		return ErrPoolClosed
-	}
-
-	p.Release()
-
-	var purgeCh <-chan struct{}
-	if !p.options.DisablePurge {
-		purgeCh = p.purgeCtx.Done()
-	} else {
-		purgeCh = p.allDone
-	}
-
-	if p.Running() == 0 {
-		p.once.Do(func() {
-			close(p.allDone)
-		})
-	}
-
-	timer := time.NewTimer(timeout)
-	defer timer.Stop()
-	for {
-		select {
-		case <-timer.C:
-			return ErrTimeout
-		case <-p.allDone:
-			<-purgeCh
-			<-p.ticktockCtx.Done()
-			if p.Running() == 0 &&
-				(p.options.DisablePurge || atomic.LoadInt32(&p.purgeDone) == 1) &&
-				atomic.LoadInt32(&p.ticktockDone) == 1 {
-				return nil
-			}
+	pool := &Pool{poolCommon: pc}
+	pool.workerCache.New = func() any {
+		return &goWorker{
+			pool: pool,
+			task: make(chan func(), workerChanCap),
 		}
 	}
-}
 
-// Reboot reboots a closed pool, it does nothing if the pool is not closed.
-// If you intend to reboot a closed pool, use ReleaseTimeout() instead of
-// Release() to ensure that all workers are stopped and resource are released
-// before rebooting, otherwise you may run into data race.
-func (p *Pool) Reboot() {
-	if atomic.CompareAndSwapInt32(&p.state, CLOSED, OPENED) {
-		atomic.StoreInt32(&p.purgeDone, 0)
-		p.goPurge()
-		atomic.StoreInt32(&p.ticktockDone, 0)
-		p.goTicktock()
-		p.allDone = make(chan struct{})
-		p.once = &sync.Once{}
-	}
-}
-
-func (p *Pool) addRunning(delta int) int {
-	return int(atomic.AddInt32(&p.running, int32(delta)))
-}
-
-func (p *Pool) addWaiting(delta int) {
-	atomic.AddInt32(&p.waiting, int32(delta))
-}
-
-// retrieveWorker returns an available worker to run the tasks.
-func (p *Pool) retrieveWorker() (w worker, err error) {
-	p.lock.Lock()
-
-retry:
-	// First try to fetch the worker from the queue.
-	if w = p.workers.detach(); w != nil {
-		p.lock.Unlock()
-		return
-	}
-
-	// If the worker queue is empty, and we don't run out of the pool capacity,
-	// then just spawn a new worker goroutine.
-	if capacity := p.Cap(); capacity == -1 || capacity > p.Running() {
-		p.lock.Unlock()
-		w = p.workerCache.Get().(*goWorker)
-		w.run()
-		return
-	}
-
-	// Bail out early if it's in nonblocking mode or the number of pending callers reaches the maximum limit value.
-	if p.options.Nonblocking || (p.options.MaxBlockingTasks != 0 && p.Waiting() >= p.options.MaxBlockingTasks) {
-		p.lock.Unlock()
-		return nil, ErrPoolOverload
-	}
-
-	// Otherwise, we'll have to keep them blocked and wait for at least one worker to be put back into pool.
-	p.addWaiting(1)
-	p.cond.Wait() // block and wait for an available worker
-	p.addWaiting(-1)
-
-	if p.IsClosed() {
-		p.lock.Unlock()
-		return nil, ErrPoolClosed
-	}
-
-	goto retry
-}
-
-// revertWorker puts a worker back into free pool, recycling the goroutines.
-func (p *Pool) revertWorker(worker *goWorker) bool {
-	if capacity := p.Cap(); (capacity > 0 && p.Running() > capacity) || p.IsClosed() {
-		p.cond.Broadcast()
-		return false
-	}
-
-	worker.lastUsed = p.nowTime()
-
-	p.lock.Lock()
-	// To avoid memory leaks, add a double check in the lock scope.
-	// Issue: https://github.com/panjf2000/ants/issues/113
-	if p.IsClosed() {
-		p.lock.Unlock()
-		return false
-	}
-	if err := p.workers.insert(worker); err != nil {
-		p.lock.Unlock()
-		return false
-	}
-	// Notify the invoker stuck in 'retrieveWorker()' of there is an available worker in the worker queue.
-	p.cond.Signal()
-	p.lock.Unlock()
-
-	return true
+	return pool, nil
 }
