+++ worker_loop_queue_test.go
@@ -1,5 +1,4 @@
 //go:build !windows
-// +build !windows
 
 package ants
